requirements.txt
Flask==2.0.1
google-cloud-speech==2.4.0
google-cloud-texttospeech==2.4.0
